remove db import config
