+++ src/main/java/com/cs544/project/ProjectApplication.java
@@ -9,13 +9,15 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.jms.annotation.EnableJms;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
 @SpringBootApplication
-@EnableJpaRepositories(value = "com.cs544.project.repositories")
+@EnableJpaRepositories(value = "com.cs544.project.repository")
 @EntityScan(value = "com.cs544.project.domain")
 @ComponentScan(basePackages = {"com.cs544.project"})
+@EnableJms
 public class ProjectApplication implements CommandLineRunner {
     private final Sender sender;
     private final DatabaseInitService dbInitService;
