--- src/main/java/com/cs544/project/repositories/CourseOfferingRepository.java
@@ -1,9 +1,0 @@
-package com.cs544.project.repositories;
-
-import com.cs544.project.domain.CourseOffering;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CourseOfferingRepository extends JpaRepository<CourseOffering, Long> {
-}
